pyproject.toml
[tool.poetry]
name = "quick-book-assistant"
version = "0.1.0"
description = "A quick booking assistant for scheduling appointments."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0"
pyyaml = "^5.4"
python-dotenv = "^0.19"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
